Reduce redundant test set up code
Make TestSetUp classes abstract since they do not contain any tests and to avoid "no runnable methods" exception

+++ src/test/java/de/uhd/ifi/se/decision/management/jira/mocks/MockAvatarManager.java
@@ -52,7 +52,7 @@ public class MockAvatarManager implements AvatarManager {
 	@Nonnull
 	@Override
 	public Avatar create(Avatar avatar) throws DataAccessException {
-		return new MockAvatar((long) 1, "Test", "Test", IconType.ISSUE_TYPE_ICON_TYPE, "SysAdmin", true);
+		return new MockAvatar(1, "Test", "Test", IconType.ISSUE_TYPE_ICON_TYPE, "SysAdmin", true);
 	}
 
 	@Nonnull
